Fixed the article/blogpost url classifier.
Internal urls on lavenir.net are not published as such anymore (I'm sure there is a very good reason for that. Really.) One week worth of articles were wrongly classified as blogposts. Good news is, they are not lost, but this will need to be addressed. In the meantime, the queuing system is fixed.

Bumped the package version.